Changes made

+++ src/app/modules/admin/admin-layout/admin-layout.component.ts
@@ -1,17 +1,22 @@
 import { Component } from '@angular/core';
 import { RouterModule } from '@angular/router';
 import { AdminService } from '../../core/services/admin.service';
+import { Router } from '@angular/router';
+import { NgIf } from '@angular/common';
+
 
 @Component({
   selector: 'app-admin-layout',
-  imports: [RouterModule],
+  imports: [RouterModule,NgIf],
   templateUrl: './admin-layout.component.html',
   styleUrl: './admin-layout.component.css'
 })
-export class AdminLayoutComponent {
+export class AdminLayoutComponent 
+{
+  
+  showLogoutAlert = false;
   username: string = 'Admin';  
-
-  constructor(private adminService: AdminService) {}
+  constructor(private adminService: AdminService,private router:Router) {}
 
   ngOnInit(): void {
     const storedName = localStorage.getItem('adminName');
@@ -31,9 +36,17 @@ export class AdminLayoutComponent {
       });
     }
   }
+ 
+logout(): void {
+  const confirmed = confirm('Are you sure you want to logout?');
 
-  logout(): void {
+  if (confirmed) {
     localStorage.clear();
-    window.location.href = '/'; 
+    this.showLogoutAlert = true;
+    setTimeout(() => {
+      this.showLogoutAlert = false;
+      this.router.navigate(['/']);
+    }, 1000);
   }
+}
 }
